+++ Reflection/test/utopia/reflection/test/swing/ScrollAreaTest.scala
@@ -38,6 +38,7 @@ object ScrollAreaTest extends App
 	// Sets up localization context
 	implicit val defaultLanguageCode: String = "EN"
 	implicit val localizer: Localizer = NoLocalization
+	implicit val context: ExecutionContext = new ThreadPool("Reflection").executionContext
 
 	// Creates the labels
 	val basicFont = Font("Arial", 12, Plain, 2)
@@ -69,7 +70,6 @@ object ScrollAreaTest extends App
 
 	// Creates the frame and displays it
 	val actionLoop = new ActorLoop(actorHandler)
-	implicit val context: ExecutionContext = new ThreadPool("Reflection").executionContext
 
 	val frame = Frame.windowed(scrollArea, "Scroll View Test", User)
 	frame.setToExitOnClose()