+++ Reflection/test/utopia/reflection/test/swing/SwitchTest.scala
@@ -34,6 +34,7 @@ object SwitchTest extends App
 	// Sets up localization context
 	implicit val defaultLanguageCode: String = "EN"
 	implicit val localizer: Localizer = NoLocalization
+	implicit val context: ExecutionContext = new ThreadPool("Reflection").executionContext
 
 	// Creates the hint labels
 	val basicFont = Font("Arial", 12, Plain, 2)
@@ -66,7 +67,6 @@ object SwitchTest extends App
 
 	// Creates the frame and displays it
 	val actionLoop = new ActorLoop(actorHandler)
-	implicit val context: ExecutionContext = new ThreadPool("Reflection").executionContext
 
 	val framing = stack.inRoundedFraming(16.any, Color.white)
 	framing.background = Color.gray(0.66)